// src/utils/types.ts
export type Currency = {
  [CurrencyEnum.USD]: number,
  [CurrencyEnum.JPY]: number,
  [CurrencyEnum.CNY]: number,
  [CurrencyEnum.CHF]: number,
  [CurrencyEnum.CAD]: number,
  [CurrencyEnum.MXN]: number,
  [CurrencyEnum.INR]: number,
  [CurrencyEnum.BRL]: number,
  [CurrencyEnum.RUB]: number,
  [CurrencyEnum.KRW]: number,
  [CurrencyEnum.IDR]: number,
  [CurrencyEnum.TRY]: number,
  [CurrencyEnum.SAR]: number,
  [CurrencyEnum.SEK]: number,
  [CurrencyEnum.NGN]: number,
  [CurrencyEnum.PLN]: number,
  [CurrencyEnum.ARS]: number,
  [CurrencyEnum.NOK]: number,
  [CurrencyEnum.TWD]: number,
  [CurrencyEnum.IRR]: number,
  [CurrencyEnum.AED]: number,
  [CurrencyEnum.COP]: number,
  [CurrencyEnum.THB]: number,
  [CurrencyEnum.ZAR]: number,
  [CurrencyEnum.DKK]: number,
  [CurrencyEnum.MYR]: number,
  [CurrencyEnum.SGD]: number,
  [CurrencyEnum.ILS]: number,
  [CurrencyEnum.HKD]: number,
  [CurrencyEnum.EGP]: number,
  [CurrencyEnum.PHP]: number,
  [CurrencyEnum.CLP]: number,
  [CurrencyEnum.PKR]: number,
  [CurrencyEnum.IQD]: number,
  [CurrencyEnum.DZD]: number,
  [CurrencyEnum.KZT]: number,
  [CurrencyEnum.QAR]: number,
  [CurrencyEnum.CZK]: number,
  [CurrencyEnum.PEN]: number,
  [CurrencyEnum.RON]: number,
  [CurrencyEnum.VND]: number,
  [CurrencyEnum.BDT]: number,
  [CurrencyEnum.HUF]: number,
  [CurrencyEnum.UAH]: number,
  [CurrencyEnum.AOA]: number,
  [CurrencyEnum.MAD]: number,
  [CurrencyEnum.OMR]: number,
  [CurrencyEnum.CUC]: number,
  [CurrencyEnum.BYR]: number,
  [CurrencyEnum.AZN]: number,
  [CurrencyEnum.LKR]: number,
  [CurrencyEnum.SDG]: number,
  [CurrencyEnum.SYP]: number,
  [CurrencyEnum.MMK]: number,
  [CurrencyEnum.DOP]: number,
  [CurrencyEnum.UZS]: number,
  [CurrencyEnum.KES]: number,
  [CurrencyEnum.GTQ]: number,
  [CurrencyEnum.URY]: number,
  [CurrencyEnum.HRV]: number,
  [CurrencyEnum.MOP]: number,
  [CurrencyEnum.ETB]: number,
  [CurrencyEnum.CRC]: number,
  [CurrencyEnum.TZS]: number,
  [CurrencyEnum.TMT]: number,
  [CurrencyEnum.TND]: number,
  [CurrencyEnum.PAB]: number,
  [CurrencyEnum.LBP]: number,
  [CurrencyEnum.RSD]: number,
  [CurrencyEnum.LYD]: number,
  [CurrencyEnum.GHS]: number,
  [CurrencyEnum.YER]: number,
  [CurrencyEnum.BOB]: number,
  [CurrencyEnum.BHD]: number,
  [CurrencyEnum.CDF]: number,
  [CurrencyEnum.PYG]: number,
  [CurrencyEnum.UGX]: number,
  [CurrencyEnum.SVC]: number,
  [CurrencyEnum.TTD]: number,
  [CurrencyEnum.AFN]: number,
  [CurrencyEnum.NPR]: number,
  [CurrencyEnum.HNL]: number,
  [CurrencyEnum.BIH]: number,
  [CurrencyEnum.BND]: number,
  [CurrencyEnum.ISK]: number,
  [CurrencyEnum.KHR]: number,
  [CurrencyEnum.GEL]: number,
  [CurrencyEnum.MZN]: number,
  [CurrencyEnum.BWP]: number,
  [CurrencyEnum.PGK]: number,
  [CurrencyEnum.JMD]: number,
  [CurrencyEnum.XAF]: number,
  [CurrencyEnum.NAD]: number,
  [CurrencyEnum.ALL]: number,
  [CurrencyEnum.SSP]: number,
  [CurrencyEnum.MUR]: number,
  [CurrencyEnum.MNT]: number,
  [CurrencyEnum.NIO]: number,
  [CurrencyEnum.LAK]: number,
  [CurrencyEnum.MKD]: number,
  [CurrencyEnum.AMD]: number,
  [CurrencyEnum.MGA]: number,
  [CurrencyEnum.XPF]: number,
  [CurrencyEnum.TJS]: number,
  [CurrencyEnum.HTG]: number,
  [CurrencyEnum.BSD]: number,
  [CurrencyEnum.MDL]: number,
  [CurrencyEnum.RWF]: number,
  [CurrencyEnum.KGS]: number,
  [CurrencyEnum.GNF]: number,
  [CurrencyEnum.SRD]: number,
  [CurrencyEnum.SLL]: number,
  [CurrencyEnum.XOF]: number,
  [CurrencyEnum.MWK]: number,
  [CurrencyEnum.FJD]: number,
  [CurrencyEnum.ERN]: number,
  [CurrencyEnum.SZL]: number,
  [CurrencyEnum.GYD]: number,
  [CurrencyEnum.BIF]: number,
  [CurrencyEnum.KYD]: number,
  [CurrencyEnum.MVR]: number,
  [CurrencyEnum.LSL]: number,
  [CurrencyEnum.LRD]: number,
  [CurrencyEnum.CVE]: number,
  [CurrencyEnum.DJF]: number,
  [CurrencyEnum.SCR]: number,
  [CurrencyEnum.SOS]: number,
  [CurrencyEnum.GMD]: number,
  [CurrencyEnum.KMF]: number,
  [CurrencyEnum.STD]: number,
  [CurrencyEnum.XRP]: number,
  [CurrencyEnum.AUD]: number,
  [CurrencyEnum.BGN]: number,
  [CurrencyEnum.BTC]: number,
  [CurrencyEnum.JOD]: number,
  [CurrencyEnum.GBP]: number,
  [CurrencyEnum.ETH]: number,
  [CurrencyEnum.EUR]: number,
  [CurrencyEnum.LTC]: number,
  [CurrencyEnum.NZD]: number
}

export type CurrencyPayload = {
  query: {
    base_currency: string,
    timestamp: number
  },
  data: Currency
}

export enum CurrencyEnum {
  USD = 'USD',
  JPY = 'JPY',
  CNY = 'CNY',
  CHF = 'CHF',
  CAD = 'CAD',
  MXN = 'MXN',
  INR = 'INR',
  BRL = 'BRL',
  RUB = 'RUB',
  KRW = 'KRW',
  IDR = 'IDR',
  TRY = 'TRY',
  SAR = 'SAR',
  SEK = 'SEK',
  NGN = 'NGN',
  PLN = 'PLN',
  ARS = 'ARS',
  NOK = 'NOK',
  TWD = 'TWD',
  IRR = 'IRR',
  AED = 'AED',
  COP = 'COP',
  THB = 'THB',
  ZAR = 'ZAR',
  DKK = 'DKK',
  MYR = 'MYR',
  SGD = 'SGD',
  ILS = 'ILS',
  HKD = 'HKD',
  EGP = 'EGP',
  PHP = 'PHP',
  CLP = 'CLP',
  PKR = 'PKR',
  IQD = 'IQD',
  DZD = 'DZD',
  KZT = 'KZT',
  QAR = 'QAR',
  CZK = 'CZK',
  PEN = 'PEN',
  RON = 'RON',
  VND = 'VND',
  BDT = 'BDT',
  HUF = 'HUF',
  UAH = 'UAH',
  AOA = 'AOA',
  MAD = 'MAD',
  OMR = 'OMR',
  CUC = 'CUC',
  BYR = 'BYR',
  AZN = 'AZN',
  LKR = 'LKR',
  SDG = 'SDG',
  SYP = 'SYP',
  MMK = 'MMK',
  DOP = 'DOP',
  UZS = 'UZS',
  KES = 'KES',
  GTQ = 'GTQ',
  URY = 'URY',
  HRV = 'HRV',
  MOP = 'MOP',
  ETB = 'ETB',
  CRC = 'CRC',
  TZS = 'TZS',
  TMT = 'TMT',
  TND = 'TND',
  PAB = 'PAB',
  LBP = 'LBP',
  RSD = 'RSD',
  LYD = 'LYD',
  GHS = 'GHS',
  YER = 'YER',
  BOB = 'BOB',
  BHD = 'BHD',
  CDF = 'CDF',
  PYG = 'PYG',
  UGX = 'UGX',
  SVC = 'SVC',
  TTD = 'TTD',
  AFN = 'AFN',
  NPR = 'NPR',
  HNL = 'HNL',
  BIH = 'BIH',
  BND = 'BND',
  ISK = 'ISK',
  KHR = 'KHR',
  GEL = 'GEL',
  MZN = 'MZN',
  BWP = 'BWP',
  PGK = 'PGK',
  JMD = 'JMD',
  XAF = 'XAF',
  NAD = 'NAD',
  ALL = 'ALL',
  SSP = 'SSP',
  MUR = 'MUR',
  MNT = 'MNT',
  NIO = 'NIO',
  LAK = 'LAK',
  MKD = 'MKD',
  AMD = 'AMD',
  MGA = 'MGA',
  XPF = 'XPF',
  TJS = 'TJS',
  HTG = 'HTG',
  BSD = 'BSD',
  MDL = 'MDL',
  RWF = 'RWF',
  KGS = 'KGS',
  GNF = 'GNF',
  SRD = 'SRD',
  SLL = 'SLL',
  XOF = 'XOF',
  MWK = 'MWK',
  FJD = 'FJD',
  ERN = 'ERN',
  SZL = 'SZL',
  GYD = 'GYD',
  BIF = 'BIF',
  KYD = 'KYD',
  MVR = 'MVR',
  LSL = 'LSL',
  LRD = 'LRD',
  CVE = 'CVE',
  DJF = 'DJF',
  SCR = 'SCR',
  SOS = 'SOS',
  GMD = 'GMD',
  KMF = 'KMF',
  STD = 'STD',
  XRP = 'XRP',
  AUD = 'AUD',
  BGN = 'BGN',
  BTC = 'BTC',
  JOD = 'JOD',
  GBP = 'GBP',
  ETH = 'ETH',
  EUR = 'EUR',
  LTC = 'LTC',
  NZD = 'NZD'
}